Berekening en UI

+++ LifePerformanceMitch/LifePerformanceMitch/Model/Huur.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -9,7 +10,7 @@ namespace LifePerformanceMitch.Model
   public interface Huur
     {
         string Naam { get; set; }
-        double Huurprijs { get; set; }
+        decimal Huurprijs { get; set; }
        
         
     }
